agents/05-scenario-modeling-agent/requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
strands-agents-tools>=0.1.0
numpy>=1.24.0
